Reorganizacion en paquetes
