+++ bin/.gitignore
@@ -1,1 +1,5 @@
 /snake/
+/frutas/
+/obstaculo/
+/logica/
+/gui/
