+++ src/gui/GameFirstClass.java
@@ -1,4 +1,4 @@
-package snake;
+package gui;
 
 import java.awt.Color;
 import java.util.ArrayList;
@@ -11,6 +11,11 @@ import java.awt.event.KeyAdapter;
 import java.awt.event.KeyEvent;
 import javax.swing.JFrame;
 import javax.swing.JPanel;
+
+import logica.Escenario;
+import snake.AL;
+import snake.Dibujable;
+
 import java.awt.BorderLayout;
 import javax.swing.JButton;
 import javax.swing.JLabel;
